requirements.txt
mkdocs>=1.6.0
mkdocs-material>=9.7.0
mkdocs-macros-plugin>=1.0.0
pyyaml>=6.0
